Moteur RPN

+++ src/main/MaxIntervalException.java
@@ -3,7 +3,7 @@ package main;
 @SuppressWarnings("serial")
 public class MaxIntervalException extends Exception {
 
-	public MaxIntervalException(int min, int max)
+	public MaxIntervalException(double min, double max)
 	{
 		super(" Le nombre entré est plus grand que la borne supérieure de l'intervalle ["+min+", "+max+"]") ;
 	}
